WIP make displayCurrencyDropdown
