+++ shared/wallets/wallet/settings/display-currency-dropdown/index.native.js
@@ -1,0 +1,3 @@
+// @flow
+import * as React from 'react'
+import type {Props} from './index.types'
