+++ shared/wallets/wallet/settings/display-currency-dropdown/index.types.js.flow
@@ -1,0 +1,10 @@
+// @flow
+import * as I from 'immutable'
+import * as Types from '../../../../constants/types/wallets'
+
+export type Props = {|
+  currencies: I.List<Types.Currency>,
+  onCurrencyChange: Types.CurrencyCode => void,
+  selected: Types.Currency,
+  waiting: boolean,
+|}
